+++ .github/CHANGELOG.md
@@ -12,6 +12,7 @@
 1. [FMGC] Don't accept blank input or / for hold distance - @tracernz (Mike)
 1. [ATSU] Fix LSK6L not returning to ATSU DATALINK page in ATC MENU - @BravoMike99 (Bruno_pt99#5802)
 1. [HYD] Trimmable physical assemblies - @Crocket63 (crocket)
+1. [HYD] Simulation of the rudder mechanical assembly and yaw dampers - @Crocket63 (crocket)
 
 ## 0.9.0